Refactor prediction hierarchy - decouple DemandPredictor (core maths) from hierarchy orchestration - rename calculation.py to calculate.py - rename orchestrator.py to orchestrate.py - move calculate_hierarchical_stats to calculate.py - rename prediction_hierarchical_level to aggregate_predictions
